1. Рефакторинг 2. Добавили Blueprints 3. Изменили инициализацию приложения
