+++ webapp/templates/create_user_page.html
@@ -37,7 +37,7 @@
         <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('create_payment') }}">Загрузка графиков</a>
+                    <a class="nav-link" href="{{ url_for('payment.create_payment') }}">Загрузка графиков</a>
                 </li>
                 <li class="nav-item dropdown">
                     <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button"
@@ -46,8 +46,8 @@
                         Информация по кредитам
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                        <li><a class="dropdown-item" href="{{ url_for('list_of_all_payments') }}">Общая таблица</a></li>
-                        <li><a class="dropdown-item" href="{{ url_for('list_of_daily_payments') }}">Мониторинг
+                        <li><a class="dropdown-item" href="{{ url_for('payment.list_of_all_payments') }}">Общая таблица</a></li>
+                        <li><a class="dropdown-item" href="{{ url_for('payment.list_of_daily_payments') }}">Мониторинг
                             платежей</a></li>
                     </ul>
                 </li>
@@ -58,15 +58,15 @@
                         Информация по пользователям
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdown">
-                        <li><a class="dropdown-item" href="{{ url_for('list_of_users') }}">Список пользователей</a></li>
-                        <li><a class="dropdown-item" href="{{ url_for('create_user') }}">Добавление нового
+                        <li><a class="dropdown-item" href="{{ url_for('user.list_of_users') }}">Список пользователей</a></li>
+                        <li><a class="dropdown-item" href="{{ url_for('user.create_user') }}">Добавление нового
                             пользователя</a></li>
                     </ul>
                 </li>
             </ul>
             <ul class="navbar-nav ms-auto">
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('exit_user') }}">Выход</a>
+                    <a class="nav-link" href="{{ url_for('user.exit_user') }}">Выход</a>
                 </li>
             </ul>
         </div>
@@ -85,7 +85,7 @@
         {% endif %}
         {% endwith %}
         <div class="card card-body">
-            <form action="{{ url_for('create_user') }}" method="POST">
+            <form action="{{ url_for('user.create_user') }}" method="POST">
                 <div class="form-group">
                     <label>Электронная почта</label>
                     <input type="text" name="type_pl" value="" class="form-control" required>
