+++ webapp/templates/login_page.html
@@ -65,7 +65,7 @@
             </div>
             {% endif %}
             {% endwith %}
-            <form action="{{ url_for('process_login') }}" method="POST">
+            <form action="{{ url_for('user.process_login') }}" method="POST">
                 {{ form.hidden_tag() }}
                 <div class="form-group">
                     {{ form.username.label }}<br>
